Switched axios to HttpClient on Angular, reorganize file structure in Angular
