+++ frontend/src/app/app.module.ts
@@ -1,17 +1,16 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
+import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ButtonsComponent } from './buttons/buttons.component';
-import { HeaderComponent } from './header/header.component';
-import { LoginFormComponent } from './login-form/login-form.component';
-import { WelcomeContentComponent } from './welcome-content/welcome-content.component';
-import { AuthContentComponent } from './auth-content/auth-content.component';
-import { ContentComponent } from './content/content.component';
-
-import { AxiosService } from './axios.service';
+import { ButtonsComponent } from './components/buttons/buttons.component';
+import { HeaderComponent } from './components/header/header.component';
+import { LoginFormComponent } from './components/login-form/login-form.component';
+import { WelcomeContentComponent } from './components/welcome-content/welcome-content.component';
+import { AuthContentComponent } from './components/auth-content/auth-content.component';
+import { ContentComponent } from './components/content/content.component';
 import { NavBarComponent } from './components/nav-bar/nav-bar.component';
 
 @NgModule({
@@ -28,9 +27,9 @@ import { NavBarComponent } from './components/nav-bar/nav-bar.component';
   imports: [
     BrowserModule,
     AppRoutingModule,
-    FormsModule
+    FormsModule,
+    HttpClientModule
   ],
-  providers: [AxiosService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
