--- frontend/src/app/auth-content/auth-content.component.ts
@@ -1,33 +1,0 @@
-import { Component } from '@angular/core';
-import { AxiosService } from '../axios.service';
-
-@Component({
-  selector: 'app-auth-content',
-  templateUrl: './auth-content.component.html',
-  styleUrls: ['./auth-content.component.css']
-})
-export class AuthContentComponent {
-  data: string[] = [];
-
-  constructor(private axiosService: AxiosService) {}
-
-  ngOnInit(): void {
-    this.axiosService.request(
-        "GET",
-        "/messages",
-        {}).then(
-        (response) => {
-            this.data = response.data;
-        }).catch(
-        (error) => {
-            if (error.response.status === 401) {
-                this.axiosService.setAuthToken(null);
-            } else {
-                this.data = error.response.code;
-            }
-
-        }
-    );
-  }
-
-}
