+++ frontend/src/app/components/auth-content/auth-content.component.spec.ts
