+++ frontend/src/app/components/welcome-content/welcome-content.component.spec.ts
