+++ frontend/src/app/http.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { AxiosService } from './axios.service';
+import { HttpService } from './http.service';
 
 describe('AxiosService', () => {
-  let service: AxiosService;
+  let service: HttpService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(AxiosService);
+    service = TestBed.inject(HttpService);
   });
 
   it('should be created', () => {
