+++ frontend/src/environments/environment.development.ts
@@ -1,0 +1,3 @@
+export const environment = {
+  baseUrl: 'http://localhost:8080',
+};
