+++ frontend/src/environments/environment.ts
@@ -1,0 +1,1 @@
+export const environment = {};
